Retain original package instead of most recent
This avoids us having to do the rename. We think keeping the original dataset is
less risky, as there could be relationships to the original package that would
be lost if removed. We're also uncertain about how complete the more recent
datasets are, possibly missing topic tags or other curated data. Instead, we'll
keep the original dataset and then update it with the usual harvest process
(outside of this deduplication process).